feat: move calendar component to ui
